# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexoraAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChatInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  sessionId?: string | null;
}

export interface SearchResult {
  title: string;
  url: string;
  snippet: string;
}

export interface ChatResponse {
  reply: string;
  sessionId: string;
  /** Where the answer came from (knowledge, example, dataset, math, search, social, command) */
  source: string;
  /** @nullable */
  searchResults?: SearchResult[] | null;
}

export interface MemoryInput {
  /** @minLength 1 */
  fact: string;
  /** @nullable */
  sessionId?: string | null;
}

export interface MemoryList {
  memories: string[];
  sessionId?: string;
}

export interface OperationResult {
  success: boolean;
  message: string;
}

export interface SuggestionList {
  suggestions: string[];
}

export type GetMemoryParams = {
sessionId?: string;
query?: string;
};

export type ClearMemoryParams = {
sessionId?: string;
};