# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NexoraAI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: nexora
    description: NexoraAI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /nexora/chat:
    post:
      operationId: sendMessage
      tags: [nexora]
      summary: Send a message to NexoraAI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /nexora/memory:
    get:
      operationId: getMemory
      tags: [nexora]
      summary: Retrieve all stored memories
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
        - name: query
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Stored memories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryList"
    delete:
      operationId: clearMemory
      tags: [nexora]
      summary: Clear all memories for a session
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Memory cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"

  /nexora/memory/remember:
    post:
      operationId: rememberFact
      tags: [nexora]
      summary: Store a fact in persistent memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "200":
          description: Fact stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"

  /nexora/suggestions:
    get:
      operationId: getSuggestions
      tags: [nexora]
      summary: Get suggested starter questions
      responses:
        "200":
          description: List of suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestionList"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
        sessionId:
          type: ["string", "null"]

    ChatResponse:
      type: object
      required: [reply, sessionId, source]
      properties:
        reply:
          type: string
        sessionId:
          type: string
        source:
          type: string
          description: Where the answer came from (knowledge, example, dataset, math, search, social, command)
        searchResults:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/SearchResult"

    SearchResult:
      type: object
      required: [title, url, snippet]
      properties:
        title:
          type: string
        url:
          type: string
        snippet:
          type: string

    MemoryInput:
      type: object
      required: [fact]
      properties:
        fact:
          type: string
          minLength: 1
        sessionId:
          type: ["string", "null"]

    MemoryList:
      type: object
      required: [memories]
      properties:
        memories:
          type: array
          items:
            type: string
        sessionId:
          type: string

    OperationResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    SuggestionList:
      type: object
      required: [suggestions]
      properties:
        suggestions:
          type: array
          items:
            type: string