# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexoraAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  sessionId?: string | null;
}