# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexoraAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResult } from './searchResult';

export interface ChatResponse {
  reply: string;
  sessionId: string;
  /** Where the answer came from (knowledge, example, dataset, math, search, social, command) */
  source: string;
  /** @nullable */
  searchResults?: SearchResult[] | null;
}