# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexoraAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  title: string;
  url: string;
  snippet: string;
}